refactor: adjust all endpoint to be set in the entrypoint of the app
